README.md: add tips and tricks
Add trick describing on using stty for debuggint our console, and
also describe how to test suspend/resume. Also document as a TODO
item we should address capturing the pts for serial console and qemu
control interface if GRUB is configured to use the console for letting
us pick what kernel to boot.

Signed-off-by: Luis R. Rodriguez <mcgrof@kernel.org>