README.md: explain lack of 'make qemuconfig' upstream
We lack 'make qemuconfig' upstream, this makes it hard to compile
fresh kernels which may have changed kconfig entries. Your compiles
may just not work.

Explain all this and why this is hard to resolve. And also hint at
the lofty goals which would enable this. For a discussion upstream
on a possible solution for many more things refer to:

https://lkml.kernel.org/r/20160818175505.GM3296@wotan.suse.de

Signed-off-by: Luis R. Rodriguez <mcgrof@kernel.org>