kvm-boot: improve MAC address conflicts
The current algorithm used for computing a MAC address
only takes into consideration the first few hex characters
and these are only used for the last 6 characters of the
MAC address. If you supply a file name with less than
6 characters we also end up using a random MAC address,
but this was just due to the lack of enough information
to create a one to one mapping of these last 6 characters.

We can fix both of these issues by using a hash of the entire
string. This way we can ensure direct mapping to the same MAC
address, regardless of how long of a filename you used for
the qcow2 image name.

Signed-off-by: Luis R. Rodriguez <mcgrof@kernel.org>