kvm-boot: add KVM_BOOT_SINGLE_GUEST and disable certain variables by default
By default kvm-boot used to set the target and linux-next target if you
did not set it. This is not safe if you are working with multiple
guests so just disable it by default, and only set them back again if
unset if you have KVM_BOOT_SINGLE_GUEST=true.

Signed-off-by: Luis R. Rodriguez <mcgrof@kernel.org>