kvm-boot: rewrite setup-kvm-switch
This rewrites setup-kvm-switch to be smart about distribution services
which may have conflicts for both vde_switch and dnsmasq. It also now
allows for the script to automatically try to fix your distribution
dnsmasq instance if it was configured to be too greedy.

To enable use on your ~/.kvmboot-setup :

export KVM_BOOT_FIX_DNSMASQ_CONFLICT="true"

If set it will detect if you have systemd, are using -7 argument for
configuration and restart the dnsmasq service for you before first
sprinkling with this for instance:

 # cat /etc/dnsmasq.d/kvm-boot
 bind-interfaces
 except-interface=tap0

You will need to add this for VPN setups if they also are using dnsmasq
and spawning their own dnsmasq instance for the VPN.

This also adds command lind parsing for a few options. We add:

	-r restart the switch and setup
	-k kill vde_switch for kvm-boot, dnsmasq for kvm-boot, and reset
	   iptables

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>